requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
aiosqlite
google-generativeai
groq
pdfplumber
pillow
python-dotenv
pydantic-settings
fuzzywuzzy
python-levenshtein
aiofiles
pandas
python-dateutil
